refactor: reorganiza pastas
