+++ src/presentation/pages/home/home.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Container } from './styles'
-import LandingPage from '../landing-page/landing'
+import LandingPage from '../landings-page/lets-party/landing'
 
 function Home() {
   return (
