+++ src/presentation/pages/landings-page/lets-party/content/section-hero.tsx
