+++ src/presentation/pages/landings-page/lets-party/content/section-packge.tsx
