+++ src/presentation/pages/landings-page/lets-party/footer/footer.tsx
