+++ src/presentation/pages/landings-page/lets-party/footer/styles.tsx
